update backend MVC structure
