+++ run.py
@@ -1,0 +1,5 @@
+from flask import *
+from app import app
+from app import route
+
+app.run(host='0.0.0.0',port=3000, debug=True)
